Add --all-targets to CI clippy and clear pre-existing test-code lints (#3830)
CI ran `cargo clippy --all -- -D warnings`, which selects all packages but
only the default (lib/bin) target set, so lints in `#[cfg(test)]` modules,
integration tests, benches, and examples could not fail the gate. Flip the
flag to `--all-targets` to close the gap, and fix the 5 pre-existing test-code
lints that flipping surfaces (all in `(lib test)` targets):

- overlay/codec.rs: identity_op (`0 | X` -> `X`)
- ledger/offer.rs: useless_vec (`vec![..]` -> `[..]`)
- herder/herder.rs: nonminimal_bool (`a == !b` -> `a != b`)
- app/lifecycle.rs: two unnecessary_cast (`batch as u64` where batch is u64)

Ordering is load-bearing: the fixes and the flag flip land together so main
never goes red. Note `.cargo/config.toml` keeps `-Aclippy::style` globally
suppressed, so this closes the target gap only; it does not make CI maximally
strict.

Refs #3809

+++ crates/app/src/app/lifecycle.rs
@@ -3514,7 +3514,7 @@ mod scp_dedup_pipeline_tests {
         // Drain a full batch of DISTINCT token-bearing SCP messages. The herder
         // is in Booting → pre-filter rejects, but the flow-control token still
         // releases at the end of pump_scp_intake (Drop on every path).
-        for slot in 0..batch as u64 {
+        for slot in 0..batch {
             let envelope = make_test_envelope(1000 + slot, 2_000_000_000);
             let mut msg = make_overlay_scp_msg(envelope);
             msg.attach_test_flow_release(Arc::clone(&flow_control), grant_tx.clone());
@@ -3534,7 +3534,7 @@ mod scp_dedup_pipeline_tests {
              exactly one SEND_MORE_EXTENDED (got {grants:?})"
         );
         assert_eq!(
-            grants[0] as u64, batch as u64,
+            grants[0] as u64, batch,
             "the grant must request the full batch of flood messages"
         );
     }
